requirements.txt
airtest==1.2.6
cached-property==1.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
decorator==5.1.1
Deprecated==1.2.13
facebook-wda==1.4.6
hrpc==1.0.9
idna==3.3
Jinja2==3.1.2
MarkupSafe==2.1.1
mss==6.1.0
numpy==1.22.4
opencv-contrib-python==4.5.5.64
Pillow==9.1.1
pocoui==1.0.87
py==1.11.0
python-xlib==0.31
pywinauto==0.6.3
requests==2.27.1
retry==0.9.2
schedule==1.1.0
six==1.16.0
urllib3==1.26.9
websocket-client==0.48.0
wrapt==1.14.1
